requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-openai
langchain-qdrant
python-dotenv